Replace tabs with spaces, seperate guidance configs
