+++ addons/missileguidance/ACE_AttackProfiles.hpp
@@ -49,20 +49,3 @@ class GVAR(AttackProfiles) {
         functionName = QFUNC(attackProfile_Ballistic);
     };
 };
-
-class GVAR(SeekerTypes) {
-    class SALH {
-        name = "";
-        visualName = "";
-        description = "";
-
-        functionName = QFUNC(seekerType_SALH);
-    };
-    class Optic {
-        name = "";
-        visualName = "";
-        description = "";
-
-        functionName = QFUNC(seekerType_Optic);
-    };
-};
