+++ addons/missileguidance/ACE_SeekerTypes.hpp
@@ -1,0 +1,16 @@
+class GVAR(SeekerTypes) {
+    class SALH {
+        name = "";
+        visualName = "";
+        description = "";
+
+        functionName = QFUNC(seekerType_SALH);
+    };
+    class Optic {
+        name = "";
+        visualName = "";
+        description = "";
+
+        functionName = QFUNC(seekerType_Optic);
+    };
+};
